+++ rtlsdr_scanner/widgets.py
@@ -151,7 +151,7 @@ class MultiButton(wx.Control):
 
         dc.DrawText(self.GetLabel(),
                     self.PADDING * 2,
-                    (rect.height - textHeight) / 2)
+                    int((rect.height - textHeight) / 2))
 
         top = (rect.height / 2) - (self.ARROW_SIZE / 4)
         bottom = top + self.ARROW_SIZE / 2
@@ -161,8 +161,8 @@ class MultiButton(wx.Control):
                         (left, top),
                         (left + self.ARROW_SIZE / 2, bottom)])
         left = right - (self.ARROW_SIZE * 2)
-        top = rect.height / 4
-        bottom = rect.height * 3 / 4
+        top = int(rect.height / 4)
+        bottom = int(rect.height * 3 / 4)
         dc.DrawLine(left, top, left, bottom)
 
     def DoGetBestSize(self):
@@ -246,7 +246,7 @@ class Led(wx.Control):
         gc.FillPath(path)
         gc.StrokePath(path)
 
-        dc.DrawText(label, height + 10, (height - textHeight) / 2)
+        dc.DrawText(label, height + 10, int((height - textHeight) / 2))
 
     def on(self, colour=wx.GREEN):
         self.timer.Stop()
